+++ packages/amplify-cli/src/domain/constants.ts
@@ -16,6 +16,7 @@ export const constants = {
   LocalNodeModules: 'cli-local-node-modules',
   ParentDirectory: 'cli-parent-directory',
   GlobalNodeModules: 'global-node-modules',
+  PackagedNodeModules: 'packaged-node-modules',
   ExecuteAmplifyCommand: 'executeAmplifyCommand',
   ExecuteAmplifyHeadlessCommand: 'executeAmplifyHeadlessCommand',
   HandleAmplifyEvent: 'handleAmplifyEvent',