+++ packages/amplify-cli/src/extensions/amplify-helpers/resource-status.ts
@@ -89,7 +89,7 @@ function getAllResources(amplifyMeta, category, resourceName, filteredResources)
   return resources;
 }
 
-function getResourcesToBeCreated(amplifyMeta, currentamplifyMeta, category, resourceName, filteredResources) {
+function getResourcesToBeCreated(amplifyMeta, currentAmplifyMeta, category, resourceName, filteredResources) {
   let resources: any[] = [];
 
   Object.keys(amplifyMeta).forEach(categoryName => {
@@ -97,9 +97,10 @@ function getResourcesToBeCreated(amplifyMeta, currentamplifyMeta, category, reso
     Object.keys(categoryItem).forEach(resource => {
       if (
         (!amplifyMeta[categoryName][resource].lastPushTimeStamp ||
-          !currentamplifyMeta[categoryName] ||
-          !currentamplifyMeta[categoryName][resource]) &&
-        categoryName !== 'providers'
+          !currentAmplifyMeta[categoryName] ||
+          !currentAmplifyMeta[categoryName][resource]) &&
+        categoryName !== 'providers' &&
+        amplifyMeta[categoryName][resource].serviceType !== 'imported'
       ) {
         amplifyMeta[categoryName][resource].resourceName = resource;
         amplifyMeta[categoryName][resource].category = categoryName;
@@ -128,9 +129,10 @@ function getResourcesToBeCreated(amplifyMeta, currentamplifyMeta, category, reso
         const dependsOnCategory = resources[i].dependsOn[j].category;
         const dependsOnResourcename = resources[i].dependsOn[j].resourceName;
         if (
-          !amplifyMeta[dependsOnCategory][dependsOnResourcename].lastPushTimeStamp ||
-          !currentamplifyMeta[dependsOnCategory] ||
-          !currentamplifyMeta[dependsOnCategory][dependsOnResourcename]
+          (!amplifyMeta[dependsOnCategory][dependsOnResourcename].lastPushTimeStamp ||
+            !currentAmplifyMeta[dependsOnCategory] ||
+            !currentAmplifyMeta[dependsOnCategory][dependsOnResourcename]) &&
+          amplifyMeta[dependsOnCategory][dependsOnResourcename].serviceType !== 'imported'
         ) {
           resources.push(amplifyMeta[dependsOnCategory][dependsOnResourcename]);
         }
@@ -141,17 +143,17 @@ function getResourcesToBeCreated(amplifyMeta, currentamplifyMeta, category, reso
   return _.uniqWith(resources, _.isEqual);
 }
 
-function getResourcesToBeDeleted(amplifyMeta, currentamplifyMeta, category, resourceName, filteredResources) {
+function getResourcesToBeDeleted(amplifyMeta, currentAmplifyMeta, category, resourceName, filteredResources) {
   let resources: any[] = [];
 
-  Object.keys(currentamplifyMeta).forEach(categoryName => {
-    const categoryItem = currentamplifyMeta[categoryName];
+  Object.keys(currentAmplifyMeta).forEach(categoryName => {
+    const categoryItem = currentAmplifyMeta[categoryName];
     Object.keys(categoryItem).forEach(resource => {
-      if (!amplifyMeta[categoryName] || !amplifyMeta[categoryName][resource]) {
-        currentamplifyMeta[categoryName][resource].resourceName = resource;
-        currentamplifyMeta[categoryName][resource].category = categoryName;
+      if ((!amplifyMeta[categoryName] || !amplifyMeta[categoryName][resource]) && categoryItem[resource].serviceType !== 'imported') {
+        currentAmplifyMeta[categoryName][resource].resourceName = resource;
+        currentAmplifyMeta[categoryName][resource].category = categoryName;
 
-        resources.push(currentamplifyMeta[categoryName][resource]);
+        resources.push(currentAmplifyMeta[categoryName][resource]);
       }
     });
   });
@@ -171,30 +173,119 @@ function getResourcesToBeDeleted(amplifyMeta, currentamplifyMeta, category, reso
   return resources;
 }
 
-async function getResourcesToBeUpdated(amplifyMeta, currentamplifyMeta, category, resourceName, filteredResources) {
+async function getResourcesToBeUpdated(amplifyMeta, currentAmplifyMeta, category, resourceName, filteredResources) {
   let resources: any[] = [];
 
   await asyncForEach(Object.keys(amplifyMeta), async categoryName => {
     const categoryItem = amplifyMeta[categoryName];
     await asyncForEach(Object.keys(categoryItem), async resource => {
-      if (currentamplifyMeta[categoryName]) {
-        if (currentamplifyMeta[categoryName][resource] !== undefined && amplifyMeta[categoryName][resource] !== undefined) {
-          const isLambdaLayer = amplifyMeta[categoryName][resource].service === FunctionServiceName.LambdaLayer;
-          const backendModified = await isBackendDirModifiedSinceLastPush(
-            resource,
-            categoryName,
-            currentamplifyMeta[categoryName][resource].lastPushTimeStamp,
-            isLambdaLayer,
-          );
+      if (
+        currentAmplifyMeta[categoryName] &&
+        currentAmplifyMeta[categoryName][resource] !== undefined &&
+        amplifyMeta[categoryName] &&
+        amplifyMeta[categoryName][resource] !== undefined &&
+        amplifyMeta[categoryName][resource].serviceType !== 'imported'
+      ) {
+        const isLambdaLayer = amplifyMeta[categoryName][resource].service === FunctionServiceName.LambdaLayer;
+        const backendModified = await isBackendDirModifiedSinceLastPush(
+          resource,
+          categoryName,
+          currentAmplifyMeta[categoryName][resource].lastPushTimeStamp,
+          isLambdaLayer,
+        );
 
-          if (backendModified) {
-            amplifyMeta[categoryName][resource].resourceName = resource;
-            amplifyMeta[categoryName][resource].category = categoryName;
-            resources.push(amplifyMeta[categoryName][resource]);
-          }
+        if (backendModified) {
+          amplifyMeta[categoryName][resource].resourceName = resource;
+          amplifyMeta[categoryName][resource].category = categoryName;
+          resources.push(amplifyMeta[categoryName][resource]);
         }
       }
     });
+  });
+
+  resources = filterResources(resources, filteredResources);
+
+  if (category !== undefined && resourceName !== undefined) {
+    resources = resources.filter(resource => resource.category === category && resource.resourceName === resourceName);
+  }
+
+  if (category !== undefined && !resourceName) {
+    resources = resources.filter(resource => resource.category === category);
+  }
+
+  return resources;
+}
+
+function getResourcesToBeSynced(amplifyMeta, currentAmplifyMeta, category, resourceName, filteredResources) {
+  let resources: any[] = [];
+
+  // For imported resource we are handling add/remove/delete in one place, because
+  // it does not involve CFN operations we still need a way to enforce the CLI
+  // to show changes status when imported resources are added or removed
+
+  Object.keys(amplifyMeta).forEach(categoryName => {
+    const categoryItem = amplifyMeta[categoryName];
+
+    Object.keys(categoryItem)
+      .filter(resource => categoryItem[resource].serviceType === 'imported')
+      .forEach(resource => {
+        // Added
+        if (
+          _.get(currentAmplifyMeta, [categoryName, resource], undefined) === undefined &&
+          _.get(amplifyMeta, [categoryName, resource], undefined) !== undefined
+        ) {
+          amplifyMeta[categoryName][resource].resourceName = resource;
+          amplifyMeta[categoryName][resource].category = categoryName;
+          amplifyMeta[categoryName][resource].sync = 'import';
+
+          resources.push(amplifyMeta[categoryName][resource]);
+        } else if (
+          _.get(currentAmplifyMeta, [categoryName, resource], undefined) !== undefined &&
+          _.get(amplifyMeta, [categoryName, resource], undefined) === undefined
+        ) {
+          // Removed
+          amplifyMeta[categoryName][resource].resourceName = resource;
+          amplifyMeta[categoryName][resource].category = categoryName;
+          amplifyMeta[categoryName][resource].sync = 'unlink';
+
+          resources.push(amplifyMeta[categoryName][resource]);
+        } else if (
+          _.get(currentAmplifyMeta, [categoryName, resource], undefined) !== undefined &&
+          _.get(amplifyMeta, [categoryName, resource], undefined) !== undefined
+        ) {
+          // Refresh - for resources that are already present, it is possible that secrets needed to be
+          // regenerated or any other data needs to be refreshed, it is a special state for imported resources
+          // and only need to be handled in env add, but no different status is being printed in status
+          amplifyMeta[categoryName][resource].resourceName = resource;
+          amplifyMeta[categoryName][resource].category = categoryName;
+          amplifyMeta[categoryName][resource].sync = 'refresh';
+
+          resources.push(amplifyMeta[categoryName][resource]);
+        }
+      });
+  });
+
+  // For remove it is possible that the the object key for the category not present in the meta so an extra iteration needed on
+  // currentAmplifyMeta keys as well
+
+  Object.keys(currentAmplifyMeta).forEach(categoryName => {
+    const categoryItem = currentAmplifyMeta[categoryName];
+
+    Object.keys(categoryItem)
+      .filter(resource => categoryItem[resource].serviceType === 'imported')
+      .forEach(resource => {
+        // Removed
+        if (
+          _.get(currentAmplifyMeta, [categoryName, resource], undefined) !== undefined &&
+          _.get(amplifyMeta, [categoryName, resource], undefined) === undefined
+        ) {
+          currentAmplifyMeta[categoryName][resource].resourceName = resource;
+          currentAmplifyMeta[categoryName][resource].category = categoryName;
+          currentAmplifyMeta[categoryName][resource].sync = 'unlink';
+
+          resources.push(currentAmplifyMeta[categoryName][resource]);
+        }
+      });
   });
 
   resources = filterResources(resources, filteredResources);
@@ -219,11 +310,11 @@ async function asyncForEach(array, callback) {
 export async function getResourceStatus(category?, resourceName?, providerName?, filteredResources?) {
   const amplifyProjectInitStatus = getCloudInitStatus();
   let amplifyMeta: $TSAny;
-  let currentamplifyMeta: $TSMeta = {};
+  let currentAmplifyMeta: $TSMeta = {};
 
   if (amplifyProjectInitStatus === CLOUD_INITIALIZED) {
     amplifyMeta = stateManager.getMeta();
-    currentamplifyMeta = stateManager.getCurrentMeta();
+    currentAmplifyMeta = stateManager.getCurrentMeta();
   } else if (amplifyProjectInitStatus === CLOUD_NOT_INITIALIZED) {
     amplifyMeta = stateManager.getBackendConfig();
   } else {
@@ -237,9 +328,10 @@ export async function getResourceStatus(category?, resourceName?, providerName?,
     throw error;
   }
 
-  let resourcesToBeCreated: any = getResourcesToBeCreated(amplifyMeta, currentamplifyMeta, category, resourceName, filteredResources);
-  let resourcesToBeUpdated: any = await getResourcesToBeUpdated(amplifyMeta, currentamplifyMeta, category, resourceName, filteredResources);
-  let resourcesToBeDeleted: any = getResourcesToBeDeleted(amplifyMeta, currentamplifyMeta, category, resourceName, filteredResources);
+  let resourcesToBeCreated: any = getResourcesToBeCreated(amplifyMeta, currentAmplifyMeta, category, resourceName, filteredResources);
+  let resourcesToBeUpdated: any = await getResourcesToBeUpdated(amplifyMeta, currentAmplifyMeta, category, resourceName, filteredResources);
+  let resourcesToBeSynced: any = getResourcesToBeSynced(amplifyMeta, currentAmplifyMeta, category, resourceName, filteredResources);
+  let resourcesToBeDeleted: any = getResourcesToBeDeleted(amplifyMeta, currentAmplifyMeta, category, resourceName, filteredResources);
 
   let allResources: any = getAllResources(amplifyMeta, category, resourceName, filteredResources);
 
@@ -248,17 +340,21 @@ export async function getResourceStatus(category?, resourceName?, providerName?,
   if (providerName) {
     resourcesToBeCreated = resourcesToBeCreated.filter(resource => resource.providerPlugin === providerName);
     resourcesToBeUpdated = resourcesToBeUpdated.filter(resource => resource.providerPlugin === providerName);
+    resourcesToBeSynced = resourcesToBeSynced.filter(resource => resource.providerPlugin === providerName);
     resourcesToBeDeleted = resourcesToBeDeleted.filter(resource => resource.providerPlugin === providerName);
     allResources = allResources.filter(resource => resource.providerPlugin === providerName);
   }
   let tagsUpdated = compareTags(stateManager.getProjectTags(), stateManager.getCurrentProjectTags());
+
   // if tags updated but no resource to apply tags, ignore tags updated
   if (allResources.filter(resource => resource.category === 'provider').length === 0) {
     tagsUpdated = false;
   }
+
   return {
     resourcesToBeCreated,
     resourcesToBeUpdated,
+    resourcesToBeSynced,
     resourcesToBeDeleted,
     tagsUpdated,
     allResources,
@@ -282,6 +378,7 @@ function compareTags(tags: Tag[], currenTags: Tag[]): boolean {
 
 export async function showResourceTable(category, resourceName, filteredResources) {
   const amplifyProjectInitStatus = getCloudInitStatus();
+
   if (amplifyProjectInitStatus === CLOUD_INITIALIZED) {
     const { envName } = getEnvInfo();
 
@@ -290,21 +387,30 @@ export async function showResourceTable(category, resourceName, filteredResource
     print.info('');
   }
 
-  const { resourcesToBeCreated, resourcesToBeUpdated, resourcesToBeDeleted, allResources, tagsUpdated } = await getResourceStatus(
-    category,
-    resourceName,
-    undefined,
-    filteredResources,
-  );
+  const {
+    resourcesToBeCreated,
+    resourcesToBeUpdated,
+    resourcesToBeDeleted,
+    resourcesToBeSynced,
+    allResources,
+    tagsUpdated,
+  } = await getResourceStatus(category, resourceName, undefined, filteredResources);
 
-  let noChangeResources = _.differenceWith(allResources, resourcesToBeCreated.concat(resourcesToBeUpdated), _.isEqual);
+  let noChangeResources = _.differenceWith(
+    allResources,
+    resourcesToBeCreated.concat(resourcesToBeUpdated).concat(resourcesToBeSynced),
+    _.isEqual,
+  );
   noChangeResources = noChangeResources.filter(resource => resource.category !== 'providers');
 
   const createOperationLabel = 'Create';
   const updateOperationLabel = 'Update';
   const deleteOperationLabel = 'Delete';
+  const importOperationLabel = 'Import';
+  const unlinkOperationLabel = 'Unlink';
   const noOperationLabel = 'No Change';
   const tableOptions = [['Category', 'Resource name', 'Operation', 'Provider plugin']];
+
   for (let i = 0; i < resourcesToBeCreated.length; ++i) {
     tableOptions.push([
       capitalize(resourcesToBeCreated[i].category),
@@ -313,6 +419,7 @@ export async function showResourceTable(category, resourceName, filteredResource
       resourcesToBeCreated[i].providerPlugin,
     ]);
   }
+
   for (let i = 0; i < resourcesToBeUpdated.length; ++i) {
     tableOptions.push([
       capitalize(resourcesToBeUpdated[i].category),
@@ -321,6 +428,31 @@ export async function showResourceTable(category, resourceName, filteredResource
       resourcesToBeUpdated[i].providerPlugin,
     ]);
   }
+
+  for (let i = 0; i < resourcesToBeSynced.length; ++i) {
+    let operation;
+
+    switch (resourcesToBeSynced[i].sync) {
+      case 'import':
+        operation = importOperationLabel;
+        break;
+      case 'unlink':
+        operation = unlinkOperationLabel;
+        break;
+      default:
+        // including refresh
+        operation = noOperationLabel;
+        break;
+    }
+
+    tableOptions.push([
+      capitalize(resourcesToBeSynced[i].category),
+      resourcesToBeSynced[i].resourceName,
+      operation /*syncOperationLabel*/,
+      resourcesToBeSynced[i].providerPlugin,
+    ]);
+  }
+
   for (let i = 0; i < resourcesToBeDeleted.length; ++i) {
     tableOptions.push([
       capitalize(resourcesToBeDeleted[i].category),
@@ -329,6 +461,7 @@ export async function showResourceTable(category, resourceName, filteredResource
       resourcesToBeDeleted[i].providerPlugin,
     ]);
   }
+
   for (let i = 0; i < noChangeResources.length; ++i) {
     tableOptions.push([
       capitalize(noChangeResources[i].category),
@@ -341,10 +474,13 @@ export async function showResourceTable(category, resourceName, filteredResource
   const { table } = print;
 
   table(tableOptions, { format: 'markdown' });
+
   if (tagsUpdated) {
     print.info('\nTag Changes Detected');
   }
-  const resourceChanged = resourcesToBeCreated.length + resourcesToBeUpdated.length + resourcesToBeDeleted.length > 0 || tagsUpdated;
+
+  const resourceChanged =
+    resourcesToBeCreated.length + resourcesToBeUpdated.length + resourcesToBeSynced.length + resourcesToBeDeleted.length > 0 || tagsUpdated;
 
   return resourceChanged;
 }