+++ packages/graphql-auth-transformer/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.20.10](https://github.com/aws-amplify/amplify-cli/compare/graphql-auth-transformer@6.20.9...graphql-auth-transformer@6.20.10) (2020-10-27)
+
+**Note:** Version bump only for package graphql-auth-transformer
+
+
+
+
+
+## [6.20.9](https://github.com/aws-amplify/amplify-cli/compare/graphql-auth-transformer@6.20.8...graphql-auth-transformer@6.20.9) (2020-10-22)
+
+**Note:** Version bump only for package graphql-auth-transformer
+
+
+
+
+
+## [6.20.8](https://github.com/aws-amplify/amplify-cli/compare/graphql-auth-transformer@6.20.7...graphql-auth-transformer@6.20.8) (2020-10-17)
+
+**Note:** Version bump only for package graphql-auth-transformer
+
+
+
+
+
 ## [6.20.7](https://github.com/aws-amplify/amplify-cli/compare/graphql-auth-transformer@6.20.6...graphql-auth-transformer@6.20.7) (2020-10-01)
 
 **Note:** Version bump only for package graphql-auth-transformer